# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}

export type WallpaperInputCategory = typeof WallpaperInputCategory[keyof typeof WallpaperInputCategory];


export const WallpaperInputCategory = {
  AMOLED_Deep_Black: 'AMOLED Deep Black',
  Cyberpunk_Neon: 'Cyberpunk Neon',
  Minimalist: 'Minimalist',
  Anime: 'Anime',
  Nature: 'Nature',
  Abstract_3D: 'Abstract 3D',
  Live_Looping_Visuals: 'Live Looping Visuals',
} as const;

export interface WallpaperInput {
  /**
     * @minLength 1
     * @maxLength 80
     */
  title: string;
  category: WallpaperInputCategory;
  /**
     * @maxItems 12
     * @items.minLength 1
     * @items.maxLength 32
     */
  tags: string[];
  /**
     * @minLength 1
     * @maxLength 256
     */
  objectPath: string;
}

export interface WallpaperItem {
  id: string;
  title: string;
  category: string;
  tags: string[];
  imageUrl: string;
  creatorId: string;
  creatorName: string;
  createdAt: string;
}

export type WallpaperList = WallpaperItem[];

export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type LogoutBrowserSessionParams = {
returnTo?: string;
};